Incremental: wip

+++ packages/incremental/computations/mixins/cacheable.ts
@@ -1,3 +1,4 @@
+import { never } from "../../../util/miscellaneous";
 import { type ComputationDescription } from "../../incremental-lib";
 import { type ValueDefinition, HashMap, objValue } from "../../utils/hash-map";
 import { ComputationResult, ok, promiseIfOk } from "../../utils/result";
@@ -8,7 +9,8 @@ import {
 } from "../basic";
 import { SubscribableComputation } from "../mixins/subscribable";
 import { RawComputation } from "../raw";
-import { DependentComputation } from "./dependent";
+import { DependentComputation, DependentContext } from "./dependent";
+import { ParentContext } from "./parent";
 
 export interface Serializable<T> {
   getTag(value: T): CacheableTag;
@@ -26,50 +28,65 @@ export type SerializableSettings<Desc, Val> =
 
 export type CacheableTag = string;
 
-type AnySubscribableDescription = ComputationDescription<
-  RawComputation<any, any> & SubscribableComputation<any>
+type CachedGet<Desc extends SubscribableDescription<any>> = {
+  readonly kind: "get";
+  readonly desc: Desc;
+  readonly tag: CacheableTag;
+  readonly serializer: Serializable<Desc>;
+};
+
+type CachedCompute<ComputeReq> = {
+  readonly kind: "compute";
+  readonly req: ComputeReq;
+  readonly serializer: Serializable<ComputeReq>;
+};
+
+export type CachedDep = CachedGet<any> | CachedCompute<any>;
+
+type SubscribableDescription<T> = ComputationDescription<
+  RawComputation<any, T> & SubscribableComputation<T>
 >;
 
 const defaultValDef1: ValueDefinition<BasicComputationDescription<any, any>> =
   objValue;
 
-class CacheEntryBuilder<Res> {
+class CacheEntryBuilder {
   // We don't know if by any chance we requested the same dependency twice
   // and got different values (because subscribers invalidation is delayed - see SubscribableComputationMixin),
   // so we keep an array instead of a map
-  private readonly deps: [AnySubscribableDescription, CacheableTag][] = [];
+  private readonly deps: CachedDep[] = [];
 
   // Track the number of times getDependency() was called
   private requested = 0;
 
   // Prevent modifications of the dependencies array that could happen after user code executed
-  // but while checkActive() is still true
+  // but while ctx.checkActive() is still true
   // (in the case the user forgot to await upon "ctx.get()")
   private locked = false;
 
-  async getDependency<T>(
-    ctx: {
-      readonly checkActive: () => void;
-      readonly get: <T>(
-        desc: ComputationDescription<
-          RawComputation<any, T> & SubscribableComputation<T>
-        >
-      ) => Promise<ComputationResult<T>>;
-    },
-    desc: ComputationDescription<
-      RawComputation<any, T> & SubscribableComputation<T>
-    >
+  checkActive() {
+    if (this.locked) {
+      throw new Error("Computation not active");
+    }
+  }
+
+  async get<T>(
+    ctx: DependentContext,
+    desc: SubscribableDescription<T>,
+    serializer: Serializable<T> | null | undefined
   ) {
-    ctx.checkActive();
+    this.checkActive();
     this.requested++;
     const result = await ctx.get(desc);
-    if (result.ok && !this.locked) {
-      ctx.checkActive();
-      if (desc.serializer) {
-        this.deps.push([
+    if (result.ok) {
+      this.checkActive();
+      if (serializer && desc.serializer) {
+        this.deps.push({
+          kind: "get",
           desc,
-          desc.serializer.valueSerializer.getTag(result.value),
-        ]);
+          tag: serializer.getTag(result.value),
+          serializer: desc.serializer.descSerializer,
+        } satisfies CachedGet<typeof desc>);
       } else {
         // TODO warn about missing serializer
       }
@@ -77,9 +94,28 @@ class CacheEntryBuilder<Res> {
     return result;
   }
 
-  make(value: Res): CacheEntry<Res> | null {
+  compute<ComputeReq>(
+    ctx: ParentContext<ComputeReq>,
+    req: ComputeReq,
+    serializer: Serializable<ComputeReq> | null | undefined
+  ) {
+    this.checkActive();
+    this.requested++;
+    ctx.compute(req);
+    if (serializer) {
+      this.deps.push({
+        kind: "compute",
+        req,
+        serializer,
+      } satisfies CachedCompute<ComputeReq>);
+    } else {
+      // TODO warn about missing serializer
+    }
+  }
+
+  make<Res>(value: Res): CacheEntry<Res> | null {
     if (this.requested !== this.deps.length) {
-      // There is no point in storing a cache entry if someof its dependencies failed or do not have a tag
+      // There is no point in storing a cache entry if we could not save some of its dependencies
       return null;
     }
     this.locked = true;
@@ -90,7 +126,7 @@ class CacheEntryBuilder<Res> {
 class CacheEntry<Res> {
   constructor(
     readonly value: Res,
-    readonly deps: readonly [AnySubscribableDescription, CacheableTag][]
+    readonly deps: readonly CachedDep[]
   ) {}
 
   getValue() {
@@ -142,6 +178,7 @@ export interface CacheableComputation<Req, Res> {
 export class CacheableComputationMixin<Req, Res> {
   public readonly db: CacheDB;
   public readonly isCacheable: boolean;
+  private firstExec: boolean;
 
   constructor(
     public readonly source: RawComputation<any, Res> & DependentComputation,
@@ -149,6 +186,7 @@ export class CacheableComputationMixin<Req, Res> {
   ) {
     this.db = source.registry.db;
     this.isCacheable = !!desc.serializer;
+    this.firstExec = true;
   }
 
   async exec(
@@ -158,41 +196,61 @@ export class CacheableComputationMixin<Req, Res> {
     if (!this.isCacheable) {
       return baseExec(ctx);
     }
-    const currentEntry = this.db.get(this.desc);
-    if (currentEntry) {
-      const cached = currentEntry.getValue();
-      try {
-        const jobs = [];
-        for (const [depDesc, tag] of currentEntry.getDeps()) {
-          const { serializer } = depDesc;
-          if (serializer == null) throw new Error("No serializer?");
-          jobs.push(
-            ctx.get(depDesc).then(result => {
-              if (
-                !result.ok ||
-                serializer.valueSerializer.getTag(result.value) !== tag
-              ) {
-                throw new Error("Outdated");
+
+    if (this.firstExec) {
+      this.firstExec = false;
+      const currentEntry = this.db.get(this.desc);
+      if (currentEntry) {
+        const cached = currentEntry.getValue();
+        try {
+          const jobs = [];
+          for (const dep of currentEntry.getDeps()) {
+            const { serializer } = dep.desc;
+            if (serializer == null) throw new Error("No serializer");
+            switch (dep.kind) {
+              case "dep": {
+                jobs.push(
+                  ctx.get(dep.desc).then(result => {
+                    if (
+                      !result.ok ||
+                      serializer.valueSerializer.getTag(result.value) !==
+                        dep.tag
+                    ) {
+                      throw new Error("Outdated");
+                    }
+                  })
+                );
+                break;
               }
-            })
-          );
+              case "child":
+                // TODO
+                break;
+              default:
+                never(dep);
+            }
+          }
+          await Promise.all(jobs);
+          return ok(cached);
+        } catch (err) {
+          // Check we are still running
+          ctx.checkActive();
+          // Reset dependencies and cache
+          this.source.dependentMixin.invalidateRoutine();
+          this.db.delete(this.desc);
         }
-        await Promise.all(jobs);
-        return ok(cached);
-      } catch (err) {}
+      }
     }
-    // Check we are still running
-    ctx.checkActive();
-    // Reset dependencies and cache
-    this.source.dependentMixin.invalidateRoutine();
-    this.db.delete(this.desc);
+
     // New entry
-    const newEntry = new CacheEntryBuilder<Res>();
+    const newEntry = new CacheEntryBuilder(ctx);
     // Execute
     const result = await baseExec({
-      ...ctx,
-      get: desc => newEntry.getDependency(ctx, desc),
-      getOk: desc => promiseIfOk(newEntry.getDependency(ctx, desc)),
+      request: ctx.request,
+      checkActive: () => newEntry.checkActive(),
+      get: desc => newEntry.get(desc, desc.serializer?.valueSerializer),
+      getOk: desc =>
+        promiseIfOk(newEntry.get(desc, desc.serializer?.valueSerializer)),
+      fs: ctx.fs,
     });
     // On success, save result
     if (result.ok) {
@@ -204,12 +262,14 @@ export class CacheableComputationMixin<Req, Res> {
 
   invalidateRoutine() {
     if (this.isCacheable) {
+      this.firstExec = false;
       this.db.delete(this.desc);
     }
   }
 
   deleteRoutine() {
     if (this.isCacheable) {
+      this.firstExec = false;
       this.db.delete(this.desc);
     }
   }
